backend/requirements.txt
fastapi==0.100.1
uvicorn==0.23.2
sqlalchemy==2.0.19
steamsignin==1.1.1
python-dotenv==1.0.0
jwt==1.3.1
aiosqlite==0.19.0

